46 facilities were not unfillable, we dropped their numbers on insert
Stephen asked why the thin-page queue gave up on 257 facilities. The log
said the websites were operator indexes without facility data. For 46 of
them that was our own error reported as theirs.

The update compared "$3 > 0" against an untyped zero, so Postgres resolved
the parameter to integer, and every facility that published a fractional
capacity failed with "invalid input syntax for type integer": 1.2 MW,
3.75 MW, 238.7 MW. Three runs of that and the row retired as unfillable.
Those pages published exactly what we went there for, and we threw it away
at the last step.

Cast to numeric, and the 46 requeued with attempts reset and a note saying
what retired them, because leaving them at three attempts would have kept
them dead behind a fixed bug.

The rest of that queue is genuinely upstream and will be reported with
URLs rather than quietly carried: 70 rows whose website is an operator
index and needs a real facility URL, 60 whose pages publish no capacity or
floor area at all, 80 blocked or missing at HTTP level, and one with a
certificate served for the wrong host.

The lesson is the one worth keeping. A failure message written by the code
that failed will describe the world, not the code. "The website has no
facility data" and "we cannot store what the website gave us" produce the
same empty page and the same retirement, and only one of them is somebody
else's problem.

+++ twoai_thinfacility.go
@@ -147,9 +147,17 @@ func twoaiThinFillFacilities(db *sql.DB) {
 			profile["address"] = r.address
 		}
 		pj, _ := json.Marshal(profile)
+		// $3 IS A DECIMAL AND POSTGRES WAS INFERRING IT AS AN INTEGER. The
+		// comparison "$3 > 0" against an untyped 0 made the driver resolve the
+		// parameter to integer, so every facility that published a fractional
+		// capacity - 1.2 MW, 3.75 MW, 238.7 MW - failed on insert with
+		// "invalid input syntax for type integer" and was retired after three
+		// runs as unfillable. It was not unfillable. Those 46 pages published
+		// exactly what we asked for and we threw it away at the last step,
+		// then logged the loss as a property of their websites.
 		if _, err := db.Exec(`UPDATE twoai_dc_facilities
 			SET profile=$2::jsonb, status='enriched',
-			    critical_it_mw = CASE WHEN $3 > 0 THEN $3 ELSE critical_it_mw END,
+			    critical_it_mw = CASE WHEN $3::numeric > 0 THEN $3::numeric ELSE critical_it_mw END,
 			    last_seen=current_date
 			WHERE id=$1`, c.ref, string(pj), r.mw); err != nil {
 			thinAttempt(db, c.path, "db: "+err.Error())
